+++ src/managednetworkfabric/azext_managednetworkfabric/__init__.py
@@ -8,6 +8,11 @@
 from azure.cli.core import AzCommandsLoader
 from azext_managednetworkfabric._help import helps  # pylint: disable=unused-import
 
+try:
+    from azext_managednetworkfabric import _breaking_change  # noqa: F401 pylint: disable=unused-import
+except ImportError:
+    pass
+
 
 class ManagednetworkfabricCommandsLoader(AzCommandsLoader):