+++ src/containerapp/HISTORY.rst
@@ -4,6 +4,7 @@ Release History
 ===============
 upcoming
 ++++++
+* 'az containerapp env storage set': Support Azure Files authentication with an Azure Key Vault secret or managed identity
 
 1.3.0b4
 ++++++